Extract middleware types to a separate file
